requirements.txt
bitmath==1.3.3.1
certifi==2020.12.5
chardet==4.0.0
decorator==4.4.2
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
moviepy==1.0.3
numpy==1.19.5
Pillow==8.1.0
proglog==0.1.9
pygifsicle==1.0.2
python-magic-bin==0.4.14
pywin32==300
requests==2.25.1
tqdm==4.56.0
urllib3==1.26.2
watchdog==1.0.2
